add support for os-variant and the concept of flavors
Add support for setting os-variant. Also add support for setting some
default values for kickstart, location and os-variant based on a 'flavor'.
Picking a flavor sets defaults for these parameters, but they can also
be overridden on the command line. For example, you can pick the flavor
'alma9', but override the kickstart template or set a non-default location
on the command line.